fix: harden Noise executor relay
