+++ codex-rs/exec-server/src/relay_proto.rs
@@ -4,5 +4,6 @@ mod generated;
 pub(crate) use generated::RelayData;
 pub(crate) use generated::RelayHandshake;
 pub(crate) use generated::RelayMessageFrame;
+pub(crate) use generated::RelayReset;
 pub(crate) use generated::RelayResume;
 pub(crate) use generated::relay_message_frame;
